requirements.txt
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
executing==0.8.2
ipython==8.0.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
